Refactor code from the Score class
Create the Frame class delegate the calculation of the frame score.

Move and adapt the necessary methods to obtain the
score to the Frame class.
